requirements.txt
streamlit
scikit-learn
facenet-pytorch
torch
torchvision
Pillow
numpy
